+++ Scripts/Edit-LiteJsonLine.ps1
@@ -3,10 +3,14 @@
 	Edits the current editor JSON line value.
 
 .Description
-	This command is called from the editor with a JSON file and the current
-	line like `<key> : "<text>"`. The line may ends with a comma. The <text>
-	is opened in another editor. After saved changes the current line is set
-	to `<key> : "<new-text>"`, with a comma if it was there.
+	This command is called from JSON editors with the current line like:
+
+		key : "<text>"
+		key : {$date: "<text>"}
+
+	The line may ends with a comma. The <text> is opened in another editor,
+	dates are converted to local time and more friendly format. After saved
+	changes the current line is updated, with a comma if it was there.
 #>
 
 function Edit-LiteJsonLine {
@@ -17,6 +21,7 @@ function Edit-LiteJsonLine {
 		throw "Invoke this command from the editor."
 	}
 
+	# get line text
 	$text = $Editor.Line.Text.TrimEnd()
 	$isComma = $text.EndsWith(',')
 	if ($isComma) {
@@ -25,6 +30,7 @@ function Edit-LiteJsonLine {
 	$spaces = if ($text -match '^\s*') {$matches[0]}
 	$text = $text.Trim()
 
+	# make json, get key and value
 	$json = "{$text}"
 	try {
 		$doc = [Ldbc.Dictionary]::FromJson($json)
@@ -34,26 +40,45 @@ function Edit-LiteJsonLine {
 	}
 	$key = @($doc.Keys)[0]
 	$value = @($doc.Values)[0]
-	if ($value -isnot [string]) {
-		throw "Value should be string."
+
+	# make text to edit
+	if ($value -is [string]) {
+		$text = $value
+	}
+	elseif ($value -is [datetime]) {
+		$text = $value.ToString('yyyy-MM-dd HH:mm:ss')
+	}
+	else {
+		throw "Value should be string or date."
 	}
 
+	# edit text
 	$arg = New-Object FarNet.EditTextArgs -Property @{
 		Title = "Edit string value '$key'"
 		Extension = '.txt'
-		Text = $value
+		Text = $text
 	}
-	$value2 = $Far.AnyEditor.EditText($arg)
-	if ($value2 -ceq $value) {
+	$text2 = $Far.AnyEditor.EditText($arg)
+	if ($text2 -ceq $text) {
 		return
 	}
 
+	# make new value
+	if ($value -is [string]) {
+		$value2 = $text2
+	}
+	else {
+		$value2 = [datetime]$text
+	}
+
+	# make new line
 	$doc[$key] = $value2
 	$text = $spaces + ($doc.Print().Trim().TrimStart('{').TrimEnd('}').Trim())
 	if ($isComma) {
 		$text += ','
 	}
 
+	# set new line
 	$Editor.Line.Text = $text
 	$Editor.Redraw()
 }